requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
reportlab
flask
flask-cors
flask-bcrypt
PyJWT
